refs task/10413 Fix Unit Test
