+++ app/Repositories/CategoryRepository.php
@@ -7,45 +7,6 @@ use App\Http\Requests\CategoryRequest;
 
 class CategoryRepository
 {   
-    public function __construct(Category $category)
-    {
-        $this->model = $category;
-    }
-
-    /**
-     * @param array $data
-     * @return Category
-     * @throws CreateCategoryErrorException
-     */
-    public function createCategory(array $data) : Category
-    {
-        try 
-        {
-            return $this->model->create($data);
-        } 
-        catch (QueryException $e) 
-        {
-            throw new CreateCategoryErrorException($e);
-        }
-    }
-
-    /**
-     * @param array $data
-     * @return bool
-     * @throws UpdateCategoryErrorException
-     */
-    public function updateCategory(array $data) : bool
-    {
-        try 
-        {
-            return $this->model->update($data);
-        } 
-        catch (QueryException $e) 
-        {
-            throw new UpdateCategoryErrorException($e);
-        }
-    }
-    
     public function all()
     {
         return Category::paginate(config('pagination.category'));
